+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Renamed `zhinst.toolkit.nodetree.nodes.WildcardResult` to `zhinst.toolkit.nodetree.helpers.NodeDict`
 * Added `active_validation` argument to `CommandTable`. By disabling it, `CommandTable` does not actively 
   validate the inputs and therefore it improves the speed for command table creation.
+* Adapt `awg.enable_sequencer` to check the acknowledged value instead of using `wait_for_state_change`. This makes it much more stable when used with short sequences.
 
 ## Version 0.4.3
 * Fix issue that prevented correct compilation of sequences for AWG cores other than the first one.